Data.Interval

Description

Interval datatype and interval arithmetic.

Unlike the intervals package (http://hackage.haskell.org/package/intervals), this module provides both open and closed intervals and is intended to be used with Rational.

For the purpose of abstract interpretation, it might be convenient to use Lattice instance. See also lattices package (http://hackage.haskell.org/package/lattices).

Interval type

data Interval r

The intervals (i.e. connected and convex subsets) over real numbers R.

module Data.ExtendedReal

data Boundary

Boundary of an interval may be open (excluding an endpoint) or closed (including an endpoint).

Since: 2.0.0

Constructors

Open

Closed

Construction

interval

Arguments

:: Ord r

=> (Extended r, Boundary)
lower bound and whether it is included

-> (Extended r, Boundary)
upper bound and whether it is included

-> Interval r

smart constructor for Interval

(<=..<=) infix 5

Arguments

:: Ord r

=> Extended r
lower bound l

-> Extended r
upper bound u

-> Interval r

closed interval [l,u]

(<..<=) infix 5

Arguments

:: Ord r

=> Extended r
lower bound l

-> Extended r
upper bound u

-> Interval r

left-open right-closed interval (l,u]

(<=..<) infix 5

Arguments

:: Ord r

=> Extended r
lower bound l

-> Extended r
upper bound u

-> Interval r

left-closed right-open interval [l, u)

(<..<) infix 5

Arguments

:: Ord r

=> Extended r
lower bound l

-> Extended r
upper bound u

-> Interval r

open interval (l, u)

whole :: Ord r => Interval r

whole real number line (-∞, ∞)

empty :: Ord r => Interval r

empty (contradicting) interval

singleton :: Ord r => r -> Interval r

singleton set [x,x]

Query

null :: Ord r => Interval r -> Bool

Is the interval empty?

isSingleton :: Ord r => Interval r -> Bool

Is the interval single point?

Since: 2.0.0

extractSingleton :: Ord r => Interval r -> Maybe r

If the interval is a single point, return this point.

Since: 2.1.0

member :: Ord r => r -> Interval r -> Bool

Is the element in the interval?

notMember :: Ord r => r -> Interval r -> Bool

Is the element not in the interval?

isSubsetOf :: Ord r => Interval r -> Interval r -> Bool

Is this a subset? (i1 `isSubsetOf` i2) tells whether i1 is a subset of i2.

isProperSubsetOf :: Ord r => Interval r -> Interval r -> Bool

Is this a proper subset? (i.e. a subset but not equal).

isConnected :: Ord r => Interval r -> Interval r -> Bool

Does the union of two range form a connected set?

Since 1.3.0

lowerBound :: Interval r -> Extended r

Lower endpoint (i.e. greatest lower bound) of the interval.

lowerBound of the empty interval is PosInf.

lowerBound of a left unbounded interval is NegInf.

lowerBound of an interval may or may not be a member of the interval.

upperBound :: Interval r -> Extended r

Upper endpoint (i.e. least upper bound) of the interval.

upperBound of the empty interval is NegInf.

upperBound of a right unbounded interval is PosInf.

upperBound of an interval may or may not be a member of the interval.

lowerBound' :: Interval r -> (Extended r, Boundary)

Lower endpoint (i.e. greatest lower bound) of the interval, together with Boundary information. The result is convenient to use as an argument for interval.

upperBound' :: Interval r -> (Extended r, Boundary)

Upper endpoint (i.e. least upper bound) of the interval, together with Boundary information. The result is convenient to use as an argument for interval.

width :: (Num r, Ord r) => Interval r -> r

Width of a interval. Width of an unbounded interval is undefined.

Universal comparison operators

(<!) :: Ord r => Interval r -> Interval r -> Bool infix 4

For all x in X, y in Y. x < y?

(<=!) :: Ord r => Interval r -> Interval r -> Bool infix 4

For all x in X, y in Y. x <= y?

(==!) :: Ord r => Interval r -> Interval r -> Bool infix 4

For all x in X, y in Y. x == y?

(>=!) :: Ord r => Interval r -> Interval r -> Bool infix 4

For all x in X, y in Y. x >= y?

(>!) :: Ord r => Interval r -> Interval r -> Bool infix 4

For all x in X, y in Y. x > y?

(/=!) :: Ord r => Interval r -> Interval r -> Bool infix 4

For all x in X, y in Y. x /= y?

Since 1.0.1

Existential comparison operators

(<?) :: Ord r => Interval r -> Interval r -> Bool infix 4

Does there exist an x in X, y in Y such that x < y?

(<=?) :: Ord r => Interval r -> Interval r -> Bool infix 4

Does there exist an x in X, y in Y such that x <= y?

(==?) :: Ord r => Interval r -> Interval r -> Bool infix 4

Does there exist an x in X, y in Y such that x == y?

Since 1.0.0

(>=?) :: Ord r => Interval r -> Interval r -> Bool infix 4

Does there exist an x in X, y in Y such that x >= y?

(>?) :: Ord r => Interval r -> Interval r -> Bool infix 4

Does there exist an x in X, y in Y such that x > y?

(/=?) :: Ord r => Interval r -> Interval r -> Bool infix 4

Does there exist an x in X, y in Y such that x /= y?

Since 1.0.1

Existential comparison operators that produce witnesses (experimental)

(<??) :: (Real r, Fractional r) => Interval r -> Interval r -> Maybe (r, r) infix 4

Does there exist an x in X, y in Y such that x < y?

Since 1.0.0

(<=??) :: (Real r, Fractional r) => Interval r -> Interval r -> Maybe (r, r) infix 4

Does there exist an x in X, y in Y such that x <= y?

Since 1.0.0

(==??) :: (Real r, Fractional r) => Interval r -> Interval r -> Maybe (r, r) infix 4

Does there exist an x in X, y in Y such that x == y?

Since 1.0.0

(>=??) :: (Real r, Fractional r) => Interval r -> Interval r -> Maybe (r, r) infix 4

Does there exist an x in X, y in Y such that x >= y?

Since 1.0.0

(>??) :: (Real r, Fractional r) => Interval r -> Interval r -> Maybe (r, r) infix 4

Does there exist an x in X, y in Y such that x > y?

Since 1.0.0

(/=??) :: (Real r, Fractional r) => Interval r -> Interval r -> Maybe (r, r) infix 4

Does there exist an x in X, y in Y such that x /= y?

Since 1.0.1

Combine

intersection :: forall r. Ord r => Interval r -> Interval r -> Interval r

intersection of two intervals

intersections :: Ord r => [Interval r] -> Interval r

intersection of a list of intervals.

Since 0.6.0

hull :: forall r. Ord r => Interval r -> Interval r -> Interval r

convex hull of two intervals

hulls :: Ord r => [Interval r] -> Interval r

convex hull of a list of intervals.

Since 0.6.0

Map

mapMonotonic :: (Ord a, Ord b) => (a -> b) -> Interval a -> Interval b

mapMonotonic f i is the image of i under f, where f must be a strict monotone function, preserving negative and positive infinities.

Operations

pickup :: (Real r, Fractional r) => Interval r -> Maybe r

pick up an element from the interval if the interval is not empty.

simplestRationalWithin :: RealFrac r => Interval r -> Maybe Rational

simplestRationalWithin returns the simplest rational number within the interval.

A rational number y is said to be simpler than another y' if

abs (numerator y) <= abs (numerator y'), and

denominator y <= denominator y'.

(see also approxRational)

Since 0.4.0

Intervals relation

relate :: Ord r => Interval r -> Interval r -> Relation

Computes how two intervals are related according to the Relation classification

Orphan instances

(RealFrac r, Floating r) => Floating (Interval r)
When results of tan or ** do not form a connected interval, a convex hull is returned instead.

Instance details

Methods

pi :: Interval r

exp :: Interval r -> Interval r

log :: Interval r -> Interval r

sqrt :: Interval r -> Interval r

(**) :: Interval r -> Interval r -> Interval r

logBase :: Interval r -> Interval r -> Interval r

sin :: Interval r -> Interval r

cos :: Interval r -> Interval r

tan :: Interval r -> Interval r

asin :: Interval r -> Interval r

acos :: Interval r -> Interval r

atan :: Interval r -> Interval r

sinh :: Interval r -> Interval r

cosh :: Interval r -> Interval r

tanh :: Interval r -> Interval r

asinh :: Interval r -> Interval r

acosh :: Interval r -> Interval r

atanh :: Interval r -> Interval r

log1p :: Interval r -> Interval r

expm1 :: Interval r -> Interval r

log1pexp :: Interval r -> Interval r

log1mexp :: Interval r -> Interval r

(Real r, Fractional r) => Fractional (Interval r)
recip returns whole when 0 is an interior point. Otherwise recip (recip xs) equals to xs without 0.

Instance details

Methods

(/) :: Interval r -> Interval r -> Interval r

recip :: Interval r -> Interval r

fromRational :: Rational -> Interval r

(Num r, Ord r) => Num (Interval r)
When results of abs or signum do not form a connected interval, a convex hull is returned instead.

Instance details

Methods

(+) :: Interval r -> Interval r -> Interval r

(-) :: Interval r -> Interval r -> Interval r

(*) :: Interval r -> Interval r -> Interval r

negate :: Interval r -> Interval r

abs :: Interval r -> Interval r

signum :: Interval r -> Interval r

fromInteger :: Integer -> Interval r

(Ord r, Read r) => Read (Interval r)

Instance details

Methods

readsPrec :: Int -> ReadS (Interval r)

readList :: ReadS [Interval r]

readPrec :: ReadPrec (Interval r)

readListPrec :: ReadPrec [Interval r]

(Ord r, Show r) => Show (Interval r)

Instance details

Methods

showsPrec :: Int -> Interval r -> ShowS

show :: Interval r -> String

showList :: [Interval r] -> ShowS

Ord r => Lattice (Interval r)

Instance details

Methods

(\/) :: Interval r -> Interval r -> Interval r

(/\) :: Interval r -> Interval r -> Interval r

Ord r => BoundedJoinSemiLattice (Interval r)

Instance details

Methods

bottom :: Interval r

Ord r => BoundedMeetSemiLattice (Interval r)

Instance details

Methods

top :: Interval r