Data.IntervalRelation

Description

Interval relations and their algebra.

Documentation

data Relation

Describes how two intervals x and y can be related. See Allen's interval algebra and Intervals and their relations.

Constructors

Before
Any element of x is smaller than any element of y, and intervals are not connected. In other words, there exists an element that is bigger than any element of x and smaller than any element of y.

JustBefore
Any element of x is smaller than any element of y, but intervals are connected and non-empty. This implies that intersection of intervals is empty, and union is a single interval.

Overlaps
Intersection of x and y is non-empty, x start and finishes earlier than y. This implies that union is a single interval, and x finishes no earlier than y starts.

Starts
x is a proper subset of y, and they share lower bounds.

During
x is a proper subset of y, but they share neither lower nor upper bounds.

Finishes
x is a proper subset of y, and they share upper bounds.

Equal
Intervals are equal.

FinishedBy
Inverse of Finishes.

Contains
Inverse of During.

StartedBy
Inverse of Starts.

OverlappedBy
Inverse of Overlaps.

JustAfter
Inverse of JustBefore.

After
Inverse of Before.

invert :: Relation -> Relation

Inverts a relation, such that invert (relate x y) = relate y x