Data.IntervalSet

Description

Interval datatype and interval arithmetic.

IntervalSet type

data IntervalSet r

A set comprising zero or more non-empty, disconnected intervals.

Any connected intervals are merged together, and empty intervals are ignored.

module Data.ExtendedReal

Construction

whole :: Ord r => IntervalSet r

whole real number line (-∞, ∞)

empty :: Ord r => IntervalSet r

empty interval set

singleton :: Ord r => Interval r -> IntervalSet r

single interval

Query

null :: IntervalSet r -> Bool

Is the interval set empty?

member :: Ord r => r -> IntervalSet r -> Bool

Is the element in the interval set?

notMember :: Ord r => r -> IntervalSet r -> Bool

Is the element not in the interval set?

isSubsetOf :: Ord r => IntervalSet r -> IntervalSet r -> Bool

Is this a subset? (is1 `isSubsetOf` is2) tells whether is1 is a subset of is2.

isProperSubsetOf :: Ord r => IntervalSet r -> IntervalSet r -> Bool

Is this a proper subset? (i.e. a subset but not equal).

span :: Ord r => IntervalSet r -> Interval r

convex hull of a set of intervals.

Construction

complement :: Ord r => IntervalSet r -> IntervalSet r

Complement the interval set.

insert :: Ord r => Interval r -> IntervalSet r -> IntervalSet r

Insert a new interval into the interval set.

delete :: Ord r => Interval r -> IntervalSet r -> IntervalSet r

Delete an interval from the interval set.

Combine

union :: Ord r => IntervalSet r -> IntervalSet r -> IntervalSet r

union of two interval sets

unions :: Ord r => [IntervalSet r] -> IntervalSet r

union of a list of interval sets

intersection :: Ord r => IntervalSet r -> IntervalSet r -> IntervalSet r

intersection of two interval sets

intersections :: Ord r => [IntervalSet r] -> IntervalSet r

intersection of a list of interval sets

difference :: Ord r => IntervalSet r -> IntervalSet r -> IntervalSet r

difference of two interval sets

Conversion

List

fromList :: Ord r => [Interval r] -> IntervalSet r

Build a interval set from a list of intervals.

toList :: Ord r => IntervalSet r -> [Interval r]

Convert a interval set into a list of intervals.

Ordered list

toAscList :: Ord r => IntervalSet r -> [Interval r]

Convert a interval set into a list of intervals in ascending order.

toDescList :: Ord r => IntervalSet r -> [Interval r]

Convert a interval set into a list of intervals in descending order.

fromAscList :: Ord r => [Interval r] -> IntervalSet r

Build a map from an ascending list of intervals. The precondition is not checked.